Refactor provider logic to core and simplify fetch options
Separate platform-specific business logic from utils

- Create core/providers directory for GitHub/GitLab business logic
  - detector.ts: Provider detection
  - github.ts: GitHub API calls
  - gitlab.ts: GitLab API calls and project validation
  - archive.ts: Archive URL generation and branch fetching
- Keep utils as pure utility functions only (parser, download, extract)
- Remove provider branching for mode option after verifying it works on GitHub

fix #20
